Separate demo logic from library
